+++ changelog.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.4.3] - [Unreleased]
+## [0.4.3] - [2022-03-18]
 
 **Added**